system2x: More accurate timing for S86/S1/S2
as it turns out, the update rate isn't exactly 60/120 hz for these
systems.

For system86/system1, the tick rate is based on MAME's refresh rate.
Timing for these systems are all controlled by the sound bus controller,
which gets a vblank input.

For system2/21, the new tick rate is based on the digital collection CD.
The timing is controlled by the C140 but this part isn't emulated...

I will probably need to update the vgm writer to use floating point delays
as there will probably be some timing errors in logged VGMs otherwise.